refactor: 아이콘 중앙화
