+++ client/src/features/participant/ui/admin-participant-delete-modal.tsx
@@ -1,6 +1,5 @@
 import { useState } from "react";
-import { AlertTriangle } from "lucide-react";
-import { Dialog, DialogContent, DialogHeader, DialogTitle, DialogDescription, DialogFooter, Button } from "@/shared/ui";
+import { Dialog, DialogContent, DialogHeader, DialogTitle, DialogDescription, DialogFooter, Button, AlertTriangle } from "@/shared/ui";
 import type { Participant } from "@/entities/participant";
 
 interface ParticipantDeleteModalProps {
