+++ client/src/shared/ui/checkbox.tsx
@@ -2,8 +2,8 @@
 
 import * as React from "react";
 import * as CheckboxPrimitive from "@radix-ui/react-checkbox";
-import { CheckIcon } from "lucide-react";
 import { cn } from "../lib";
+import { CheckIcon } from "./icons";
 
 function Checkbox({ className, ...props }: React.ComponentProps<typeof CheckboxPrimitive.Root>) {
   return (
