+++ client/src/shared/ui/icons.ts
@@ -1,0 +1,68 @@
+// 공통으로 사용되는 아이콘들을 한 곳에서 관리
+// 트리 셰이킹 최적화와 중복 제거를 위함
+
+export {
+  // Alert & Status Icons
+  AlertCircle,
+  AlertTriangle,
+  CheckCircle,
+  XCircle,
+  Info,
+  
+  // Action Icons
+  Plus,
+  Edit,
+  Trash2,
+  Check,
+  X,
+  
+  // Navigation Icons
+  ChevronLeft,
+  ChevronRight,
+  ChevronUpIcon as ChevronUp,
+  ChevronDownIcon as ChevronDown,
+  ChevronUpIcon,
+  ChevronDownIcon,
+  
+  // Media & Control Icons
+  Play,
+  Pause,
+  Square,
+  SkipForward,
+  RotateCcw,
+  
+  // UI Icons
+  Settings,
+  Clock,
+  Timer,
+  List,
+  Link,
+  Upload,
+  
+  // User & People Icons
+  User,
+  Users,
+  UserPlus,
+  UserCog,
+  Shield,
+  
+  // Data & Charts
+  BarChart3,
+  TrendingUp,
+  Calculator,
+  FileText,
+  FileSpreadsheet,
+  
+  // Layout & Display
+  LayoutDashboard,
+  Trophy,
+  Activity,
+  Wifi,
+  
+  // Specific Icons
+  ImagePlus,
+  Loader2,
+  Minus,
+  CheckIcon,
+  XIcon,
+} from "lucide-react";
